requirements.txt
# LLM Providers
groq>=0.4.0
google-genai>=0.3.0
instructor>=1.3.0

# Structured Output
pydantic>=2.0.0

# Audio/Voice
sounddevice>=0.4.6
webrtcvad>=2.0.10
numpy>=1.24.0

# macOS Integration
pyobjc-framework-Cocoa>=9.0

# Image Processing
pillow>=10.0.0
rembg>=2.0.50

# Memory
chromadb>=0.4.0

# Utilities
python-dotenv>=1.0.0
